Fix indentation from commit 1f5eeb7be4a
This is on the way to fixing a bug that may have existed before that
commit but that for some reason only started manifesting for me now:

  Debugger entered--Lisp error: \
  (wrong-type-argument number-or-marker-p nil)
  funcall(<= nil 6)
  <=(nil 6)
  [...]
  gnus-get-unread-articles(nil nil)
  gnus-setup-news(nil nil nil)
  gnus-1(nil nil nil)
  gnus()

Because `info' is nil (and apparently has been nil since it was
created in the `let'), `(gnus-info-level info)' also returns nil,
so we end up doing a `<=' comparison between nil and 6, the latter
being the result of the inner `if' in this reindented code.

This is just a formatting fix before I make any code changes.